requirements.txt
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.8.2
pytz==2019.1
sqlparse==0.3.0
